Separate rollingupdate e2e
